Fix test for creating marketplace items
In the NFTMarketplace contract, the test for creating marketplace items was failing
due to an incorrect comparison of the expected value for the price of the item.
 The test was expecting a value of 1000000000000000000 wei,
but the actual value was not being converted correctly from the price passed as an argument to the createMarketItem function.

To fix the test, we have added a conversion to the expected unit (wei) before passing the price to the function. Additionally, we adjusted the order of the parameters in the withArgs assertion to match the expected order of the emitted event parameters.